Issue #15 copy settings from recently created project, using mocha+chai now
