+++ .eslintrc.js
@@ -4,8 +4,8 @@ module.exports = {
     node: true
   },
   'extends': [
-    'standard',
-    'plugin:vue/recommended'
+    'plugin:vue/essential',
+    'eslint:recommended'
   ],
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
