+++ cypress.json
@@ -1,0 +1,3 @@
+{
+  "pluginsFile": "tests/e2e/plugins/index.js"
+}
