--- test/unit/index.js
@@ -1,3 +1,0 @@
-// require all test files
-const testsContext = require.context('./', true, /\.test/)
-testsContext.keys().forEach(testsContext)
