+++ tests/unit/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  env: {
+    mocha: true
+  }
+}
