--- vue.config.js
@@ -1,16 +1,0 @@
-var webpack = require('webpack')
-
-module.exports = {
-  devServer: {
-    disableHostCheck: true
-  },
-  configureWebpack: {
-    plugins: [
-      new webpack.DefinePlugin({
-        'process.env': {
-          PACKAGE_JSON: '"' + escape(JSON.stringify(require('./package.json'))) + '"'
-        }
-      })
-    ]
-  }
-}
